refactor(server): extract run_generator + reset_engine to ws_session.py
The 290-line generator closure — the last whale — moves to ws_session.py
as a standalone `run_generator(conn, world_engine, session)` function.
The `_PendingFlush` dataclass and `_flush_pending` helper become module-
private; `reset_engine` is now a sync function in ws_session.py that
takes the same trio of args. `build_error_message` follows.

Two structural cleanups:
- `WorldEngineManager.init_session` is now sync (matching `reset_state`
  from step 6); the pre-init handshake call site uses `asyncio.to_thread`.
  This drops the last `run_coro` site from the generator — no more
  `asyncio.run_coroutine_threadsafe(...).result()` cross-thread waits.
- The generator's enclosing-scope dependencies were already conn methods
  or module-level functions, so the lift is mechanical (~290 lines moved
  with name swaps).

server.py drops to ~840 lines; the only remaining body inside
`websocket_endpoint` is the lifespan setup, the pre-init handshake
loop, the post-warmup setup, and the asyncio.gather wiring.

Refs #78.
